+++ cmd/deploy/swarm/one/metrics.sh
@@ -5,11 +5,17 @@ echo "  -----$1 Fix metrics------"
 echo "  ------------------------------"
 echo ""
 trap 'echo " stop" ' INT TERM
-#chmod +x droplet_metrics.sh && ./droplet_metrics.sh 1
+#chmod +x metrics.sh && ./metrics.sh 1
 
-#$1 - the IP of this node
+ip=$1   # the IP of this node
 
-ssh root@$1 "
+if [ -z "${ip}" ]; then
+    chmod +x ./../print_error.sh && ./../print_error.sh \
+        "$0 <ip of the node>"
+    exit 1
+fi
+
+ssh root@$ip "
 yes | sudo apt-get purge do-agent
 curl -sSL https://repos.insights.digitalocean.com/install.sh | sudo bash
 /opt/digitalocean/bin/do-agent --version
